add confirm dialog for deleting user

+++ src/main/webapp/WEB-INF/views/manage/userList.jsp
@@ -6,6 +6,13 @@
 <script>
 	$(document).ready(function() {
 		$("#account-tab").addClass("active");
+		$(".confirm").confirm({
+		    text: "确定要删除这个用户？",
+		    title: "请确认",
+		    
+		    confirmButton: "确认",
+		    cancelButton: "取消"
+		});
 	});
 </script>
 </head>
@@ -22,28 +29,23 @@
 		<tags:message content="${message}" />
 		<div class="row">
 			<div class="col-xs-12 col-sm-12">
-				<form:form id="searchForm" modelAttribute="user"
-					action="${ctx}/manage/userList" method="post" class="form-inline"
-					role="form">
+				<form:form id="searchForm" modelAttribute="user" action="${ctx}/manage/userList" method="post"
+					class="form-inline" role="form">
 					<div class="form-group">
-						<label for="username">登录名：</label> 
-						<form:input type="text" name="search_LIKE_username"
-						class="form-control" path="username"></form:input>
+						<label for="username">登录名：</label>
+						<form:input type="text" name="search_LIKE_username" class="form-control" path="username"></form:input>
 					</div>
 					<div class="form-group">
-						<label for="email">邮件名：</label> 
-						<form:input type="text" name="search_EQ_email" class="form-control"
-						path="email"></form:input>
+						<label for="email">邮件名：</label>
+						<form:input type="text" name="search_EQ_email" class="form-control" path="email"></form:input>
 					</div>&nbsp;&nbsp;
-					<input id="btnSubmit" class="btn btn-primary" type="submit"
-						value="查询" />
+					<input id="btnSubmit" class="btn btn-primary" type="submit" value="查询" />
 				</form:form>
 			</div>
 		</div>
 		<br>
 
-		<table id="contentTable"
-			class="table table-striped table-bordered table-condensed">
+		<table id="contentTable" class="table table-striped table-bordered table-condensed">
 			<thead>
 				<tr>
 					<th>姓名</th>
@@ -64,10 +66,8 @@
 						<td>${user.wifiStatus == 1 ?'可接入互联网':'不可使用网络'}&nbsp;</td>
 						<td>${ww:getLabel('user_type',user.userType,'') }&nbsp;</td>
 						<td>${user.createTime}&nbsp;</td>
-						<td><a href="${ctx}/manage/userForm?id=${user.id}"
-							id="editLink-${user.username}">修改</a> <a
-							href="${ctx}/manage/deleteUser/${user.id}"
-							id="editLink-${user.username}">删除</a></td>
+						<td><a href="${ctx}/manage/userForm?id=${user.id}" id="editLink-${user.username}" class="confirm" >修改</a>
+							<a href="${ctx}/manage/deleteUser/${user.id}" id="editLink-${user.username}" onclick="return confirm('确定删除该用户？')">删除</a></td>
 					</tr>
 				</c:forEach>
 			</tbody>
